Add SessionExtendRequestSubscriber, SessionExtendResponseSubscriber
