+++ Sources/WalletConnectSign/Sign/SignClient.swift
@@ -110,6 +110,8 @@ public final class SignClient: SignClientProtocol {
     private let nonControllerSessionStateMachine: NonControllerSessionStateMachine
     private let controllerSessionStateMachine: ControllerSessionStateMachine
     private let sessionExtendRequester: SessionExtendRequester
+    private let sessionExtendRequestSubscriber: SessionExtendRequestSubscriber
+    private let sessionExtendResponseSubscriber: SessionExtendResponseSubscriber
     private let appProposeService: AppProposeService
     private let historyService: HistoryService
     private let cleanupService: SignCleanupService
@@ -140,6 +142,8 @@ public final class SignClient: SignClientProtocol {
          nonControllerSessionStateMachine: NonControllerSessionStateMachine,
          controllerSessionStateMachine: ControllerSessionStateMachine,
          sessionExtendRequester: SessionExtendRequester,
+         sessionExtendRequestSubscriber: SessionExtendRequestSubscriber,
+         sessionExtendResponseSubscriber: SessionExtendResponseSubscriber,
          appProposeService: AppProposeService,
          disconnectService: DisconnectService,
          historyService: HistoryService,
@@ -155,6 +159,8 @@ public final class SignClient: SignClientProtocol {
         self.nonControllerSessionStateMachine = nonControllerSessionStateMachine
         self.controllerSessionStateMachine = controllerSessionStateMachine
         self.sessionExtendRequester = sessionExtendRequester
+        self.sessionExtendRequestSubscriber = sessionExtendRequestSubscriber
+        self.sessionExtendResponseSubscriber = sessionExtendResponseSubscriber
         self.appProposeService = appProposeService
         self.historyService = historyService
         self.cleanupService = cleanupService
@@ -402,13 +408,13 @@ public final class SignClient: SignClientProtocol {
         controllerSessionStateMachine.onNamespacesUpdate = { [unowned self] topic, namespaces in
             sessionUpdatePublisherSubject.send((topic, namespaces))
         }
-        controllerSessionStateMachine.onExtend = { [unowned self] topic, date in
-            sessionExtendPublisherSubject.send((topic, date))
-        }
         nonControllerSessionStateMachine.onNamespacesUpdate = { [unowned self] topic, namespaces in
             sessionUpdatePublisherSubject.send((topic, namespaces))
         }
-        nonControllerSessionStateMachine.onExtend = { [unowned self] topic, date in
+        sessionExtendRequestSubscriber.onExtend = { [unowned self] topic, date in
+            sessionExtendPublisherSubject.send((topic, date))
+        }
+        sessionExtendResponseSubscriber.onExtend = { [unowned self] topic, date in
             sessionExtendPublisherSubject.send((topic, date))
         }
         sessionEngine.onEventReceived = { [unowned self] topic, event, chainId in
