+++ Sources/WalletConnectSign/Sign/SignClientFactory.swift
@@ -31,6 +31,8 @@ public struct SignClientFactory {
         let nonControllerSessionStateMachine = NonControllerSessionStateMachine(networkingInteractor: networkingClient, kms: kms, sessionStore: sessionStore, logger: logger)
         let controllerSessionStateMachine = ControllerSessionStateMachine(networkingInteractor: networkingClient, kms: kms, sessionStore: sessionStore, logger: logger)
         let sessionExtendRequester = SessionExtendRequester(sessionStore: sessionStore, networkingInteractor: networkingClient)
+        let sessionExtendRequestSubscriber = SessionExtendRequestSubscriber(networkingInteractor: networkingClient, sessionStore: sessionStore, logger: logger)
+        let sessionExtendResponseSubscriber = SessionExtendResponseSubscriber(networkingInteractor: networkingClient, sessionStore: sessionStore, logger: logger)
         let sessionTopicToProposal = CodableStore<Session.Proposal>(defaults: RuntimeKeyValueStorage(), identifier: SignStorageIdentifiers.sessionTopicToProposal.rawValue)
         let approveEngine = ApproveEngine(
             networkingInteractor: networkingClient,
@@ -62,6 +64,8 @@ public struct SignClientFactory {
             nonControllerSessionStateMachine: nonControllerSessionStateMachine,
             controllerSessionStateMachine: controllerSessionStateMachine,
             sessionExtendRequester: sessionExtendRequester,
+            sessionExtendRequestSubscriber: sessionExtendRequestSubscriber,
+            sessionExtendResponseSubscriber: sessionExtendResponseSubscriber,
             appProposeService: appProposerService,
             disconnectService: disconnectService,
             historyService: historyService,
